fix: improved server managment
