+++ pkg/server/types.go
@@ -4,10 +4,18 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/rookout/piper/pkg/clients"
 	"github.com/rookout/piper/pkg/conf"
+	"net/http"
 )
 
 type Server struct {
 	router  *gin.Engine
 	config  *conf.GlobalConfig
 	clients *clients.Clients
+}
+
+type Interface interface {
+	startServer() *http.Server
+	registerMiddlewares()
+	getRoutes()
+	ListenAndServe() *http.Server
 }
